intro Model trait and refactor
Added a Model trait that provides interfaces for reading data as a
Polars dataframe and spliting the data into X and y training data.

Renamed the datasets directory to openml since most of it was focused
on reading openml data (some further refactoring to move the code that
reads from file out but that is for another day).

Added a new datasets directory that is focused on highlevel datasets,
e.g., minst (petals) and xor.

Signed-off-by: Syd Logan <slogan621@gmail.com>